AppRTCDemoActivity: use differnet Themes for different API levels
The current AndroidManifest.xml hardcodes a Theme that
is only available in Android L or later (Material). To
maintain backwards compatibility, and for better App
style, create a single Theme/Style and define it for
different APIs.

I tested this in two Nexus %, one with prerelease L
and another with a KK 4.4.2 and the Theme is indeed
automagically updated :)

Note that this is compatible with
https://webrtc-codereview.appspot.com/26979004/

R=glaznev@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/26019004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7619 4adac7df-926f-26a2-2b94-8c16560cd09d